backend/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.1.0
apt-xapian-index==0.49
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Authlib==1.3.0
awscli==1.22.34
Babel==2.13.1
bcrypt==3.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
botocore==1.23.34
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.4
comm==0.2.0
command-not-found==0.3
confection==0.1.4
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
cymem==2.0.8
data==0.4
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-rest-auth==6.0.0
Django==5.0.4
django-allauth==0.61.1
django-cors-headers==4.4.0
django-mysql==4.14.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.17.1
duplicity==0.8.21
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
Faker==30.1.0
fasteners==0.14.1
fastjsonschema==2.19.0
filelock==3.13.1
Flask==3.0.3
Flask-Mail==0.9.1
flatbuffers==23.5.26
fonttools==4.47.2
fqdn==1.5.1
fsspec==2024.9.0
funcsigs==1.0.2
future==0.18.2
gast==0.5.4
gensim==4.3.2
google-api-core==2.19.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.63.2
gpg==1.16.0
grpcio==1.66.0
grpcio-status==1.66.0
gyp==0.1
h11==0.14.0
h5py==3.10.0
httplib2==0.20.2
huggingface-hub==0.24.6
idna==3.3
importlib-metadata==4.6.4
iniconfig==1.1.1
ipykernel==6.27.1
ipython==8.19.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.9.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.1
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.11.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.15.0
keyring==23.5.0
kiwisolver==1.4.5
langcodes==3.3.0
language-selector==0.1
latex==0.7.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.1.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
ml-dtypes==0.2.0
mlxtend==0.23.1
monotonic==1.6
more-itertools==8.10.0
murmurhash==1.0.10
mutagen==1.45.1
mysqlclient==2.1.1
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.0
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==3.11.0
plotly==5.18.0
pluggy==0.13.0
preshed==3.0.9
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==4.25.4
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
PySocks==1.7.1
pytest==6.2.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.1
regex==2023.10.3
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roman==3.3
rpds-py==0.13.2
rsa==4.9
s3transfer==0.5.0
safetensors==0.4.5
scikit-learn==1.4.0
scipy==1.11.4
seaborn==0.13.1
SecretStorage==3.3.1
selenium==4.19.0
Send2Trash==1.8.2
shutilwhich==1.1.0
simplejson==3.17.6
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==2.4.8
stack-data==0.6.3
systemd-python==234
tempdir==0.7.1
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminado==0.18.0
textblob==0.18.0.post0
thinc==8.2.3
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.44.2
trio==0.25.0
trio-websocket==0.11.1
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.10.0
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.24.6
wadllib==1.3.6
wasabi==1.1.2
wcwidth==0.2.12
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==9.1
Werkzeug==3.0.1
Whoosh==2.7.4
widgetsnbextension==4.0.9
wordcloud==1.9.3
wrapt==1.14.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
yt-dlp==2022.4.8
zipp==1.0.0
